Added coherent race type of interaction
Added coherent race type of interaction.
After a photon travels a certain distance and experiences an interaction. That type of interaction can be coherent race.
In order to calculate the polar angle of difusion when a photon experiences coherent race, a new class needed to be added.
DistributionFunctionTableProvider consists of operations needed to calculate polar angle of difusion.